test: isolate enrichment helper tests
